+++ app/index.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const {ipcRenderer} = require('electron');
 const {dialog} = require('electron').remote;
 
@@ -12,8 +10,8 @@ document.addEventListener('DOMContentLoaded', function() {
   document.getElementById('directory_select').addEventListener('click', function() {
     dialog.showOpenDialog({
       properties: ['openDirectory'],
-    }, function(path) {
-      document.getElementById('directory').value = path[0];
+    }).then((result) => {
+      document.getElementById('directory').value = result.filePaths[0];
     });
   });