+++ main.js
@@ -23,6 +23,10 @@ app.on('ready', function() {
   ejsElectron.data('version', packageJson.version);
 
   mainWindow = new BrowserWindow({
+    webPreferences: {
+      nodeIntegration: true,
+      enableRemoteModule: true,
+    },
     width: 800,
     height: 600,
   });